[REFACTOR] Put main and tests sources into separate folders on root level
The less folder hierarchy the better. We aknowledge that unit tests files are technically “source files” to, but these are special kind of source files, so we put main source files (which are the library/pod actual functionality that supposed to be used by 3rd party directly) into ‘Sources’ folder, and unit tests source files we put into special ‘Tests’ folder. That simplifies folder structure and increases discoverability (before that change a quick look on repo may convince unfamiliar person that this repo has no tests, but now it will be easy to spot the ‘Tests’ folder right at the repo root level).